[core] [同步代码]修复A3 场景register 失败问题



# message auto-generated for no-merge-commit merge:
!672 merge develop_0514_sync_register into develop

[core] [同步代码]修复A3 场景register 失败问题

Created-by: p3rry
Commit-by: p3rry
Merged-by: yrewzjsx
Description: # Pull Request Template

## Description

<!-- Provide a clear summary of the change, the problem it solves, and the technical approach. -->

## Related Issues

<!-- Replace with actual issue numbers. Example: Closes #123, Related to #456 -->
https://gitcode.com/Ascend/memfabric_hybrid/issues/122

## Type of Change

- [ ] Bug fix
- [ ] New feature
- [ ] Performance optimization
- [ ] Documentation update
- [ ] Other (please describe):_______________

## Testing

<!-- Describe how the changes were tested. Include:
- Unit/integration test coverage
- Manual validation steps
- Links to CI reports (e.g., CIDA) if available
- Note any pending tests
-->

## Checklist

- [ ] I have performed a self-review of my own code.  
- [ ] I have updated the documentation.  
- [ ] I have added tests to prove my changes are effective.

See merge request: Ascend/memfabric_hybrid!672

+++ src/hybm/csrc/mm/hybm_conn_based_segment.h
@@ -76,8 +76,8 @@ private:
     uint64_t totalVirtualSize_{0UL};
     uint8_t *localVirtualBase_{nullptr};
     uint64_t allocatedSize_{0UL};
-    std::map<uint16_t, MemSliceStatus> slices_;
-    std::map<uint16_t, std::string> exportMap_;
+    std::map<uint32_t, MemSliceStatus> slices_;
+    std::map<uint32_t, std::string> exportMap_;
     std::vector<HostExportInfo> imports_;
     std::set<uint64_t> mappedGvaMem_; // gva sets
 };
